Replace opensuse 13 with Leap
Suse 13 is EOL. So much so that the first version of leap is actually EOL as well. It's way dead.

Instead use leap and make sure we have everything we need for init scripts. That's aaa_base and the silly perl getopt-long-descriptive module which should be a dep there, but isn't.

Signed-off-by: Tim Smith <tsmith@chef.io>