Remove optional systemd services
This removes all the unnecessary services that are enabled by default and are
not needed for running inside of containers. The only service that should be
running is journald.

This also seems to fix issues noticed running these images on Linux based
systems where the system either pauses for a few seconds while the container is
"booting" or experience logout issues.

In addition, this also speeds up container "booting" since it only needs to run
journald on start up.

Some other notes:

- Disable exim service on Debian
- Disable ondemand and apparmor services on Ubuntu

Signed-off-by: Lance Albertson <lance@osuosl.org>