Selectively disable optional services to improve usage on Linux hosts
Takes a more selective approach to resolving the issues some users are facing
using these images on Linux hosts. The primary issue seems to be the fact that
all of the udev services are causing issues and should be disabled on startup.

For the use case of these containers, needing to manage user space devices isn't
needed so this should be safe to remove.

In addition, this removes or disables other services that are causing problems
and also ensures that all the services are booting up properly.

Signed-off-by: Lance Albertson <lance@osuosl.org>