Mask systemd-journal-flush.service on RHEL >=8 & Fedora
We started running into issues similar to the following [1] where running yum
commands would fail and were similar to the problems mentioned here [2]. After
doing some digging, I noticed that systemctl status was still initializating on
the centos-stream-8 image and it was due to the systemd-journal-flush.service
service.

We shouldn't need this for dokken so let's just mask this for any RHEL-based
system 8 or newer along with all Fedora. I'm not sure if this is going to be
needed elsewhere but this is a start.

[1] https://github.com/sous-chefs/yum/actions/runs/4452618323/jobs/7820436622?pr=206#step:4:222
[2] https://www.jeffgeerling.com/blog/2020/resolving-fedora-dnf-error-no-such-file-or-directory-varlibdnfrpmdblockpid

Signed-off-by: Lance Albertson <lance@osuosl.org>